Split out legacy import path from realm manager
